chore: fix wrong file names
